Implement user authentication and remove debug messages
This update integrates user authentication functionality into the apiUsers. It removes redundant log statements across different parts of the codebase and introduces the SignUp and Login pages. In these pages, there is a global error state management mechanism and redirection functionality for users after they log in. The changes also include modifications in the proxy path and the creation of a 'signup' route.